app/requirements.txt
argcomplete==3.2.2
asgiref==3.7.2
<<<<<<< HEAD
<<<<<<< HEAD
=======
>>>>>>> 1e155a63d7b775eeaf726c2d137dd5d6d9ef8a48
attrs==23.2.0
awsebcli==3.20.10
beautifulsoup4==4.12.3
boto3==1.34.39
botocore==1.34.39
build==1.0.3
cement==3.0.8
certifi==2024.2.2
<<<<<<< HEAD
=======
attrs==23.1.0
beautifulsoup4==4.12.2
boto3==1.29.3
botocore==1.32.3
certifi==2023.11.17
>>>>>>> 466e9623268ee4875868b6b8d9ca6a67e70f5d8a
=======
>>>>>>> 1e155a63d7b775eeaf726c2d137dd5d6d9ef8a48
cffi==1.16.0
cfn-flip==1.3.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
<<<<<<< HEAD
<<<<<<< HEAD
distlib==0.3.8
=======
>>>>>>> 466e9623268ee4875868b6b8d9ca6a67e70f5d8a
djangify==1.0.0
Django==5.0.2
django-bootstrap-v5==1.0.11
<<<<<<< HEAD
=======
distlib==0.3.8
djangify==1.0.0
Django==5.0.2
django-bootstrap-v5==1.0.11
>>>>>>> 1e155a63d7b775eeaf726c2d137dd5d6d9ef8a48
django-environ==0.11.2
django-storages==1.14.2
durationpy==0.6
environ==1.0
exceptiongroup==1.2.0
filelock==3.13.1
gunicorn==21.2.0
=======
durationpy==0.5
exceptiongroup==1.1.3
>>>>>>> 466e9623268ee4875868b6b8d9ca6a67e70f5d8a
h11==0.14.0
hjson==3.1.0
idna==3.6
jmespath==1.0.1
kappa==0.6.0
MarkupSafe==2.1.5
numpy==1.26.4
opencv-python==4.9.0.80
outcome==1.3.0.post0
packaging==23.2
<<<<<<< HEAD
<<<<<<< HEAD
=======
>>>>>>> 1e155a63d7b775eeaf726c2d137dd5d6d9ef8a48
pathspec==0.12.1
pep517==0.13.1
pillow==10.2.0
pip-tools==7.3.0
placebo==0.9.0
platformdirs==4.2.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycparser==2.21
pyee==11.1.0
pypi==2.1
pypiwin32==223
pyproject_hooks==1.0.0
<<<<<<< HEAD
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-ffmpeg==2.0.10
python-slugify==8.0.4
pywin32==306
PyYAML==6.0.1
requests==2.31.0
s3transfer==0.10.0
scipy==1.12.0
selenium==4.17.2
semantic-version==2.10.0
setuptools==68.2.2
=======
Pillow==10.1.0
placebo==0.9.0
pycparser==2.21
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-slugify==8.0.1
PyYAML==6.0.1
requests==2.31.0
s3transfer==0.7.0
selenium==4.15.2
>>>>>>> 466e9623268ee4875868b6b8d9ca6a67e70f5d8a
=======
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-ffmpeg==2.0.10
python-slugify==8.0.4
pywin32==306
PyYAML==6.0.1
requests==2.31.0
s3transfer==0.10.0
scipy==1.12.0
selenium==4.17.2
semantic-version==2.10.0
setuptools==68.2.2
>>>>>>> 1e155a63d7b775eeaf726c2d137dd5d6d9ef8a48
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
<<<<<<< HEAD
<<<<<<< HEAD
=======
>>>>>>> 1e155a63d7b775eeaf726c2d137dd5d6d9ef8a48
termcolor==2.4.0
text-unidecode==1.3
toml==0.10.2
tomli==2.0.1
tqdm==4.66.2
trio==0.24.0
<<<<<<< HEAD
trio-websocket==0.11.1
troposphere==4.6.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.2.0
virtualenv==20.25.0
wcwidth==0.2.13
webdriver-manager==4.0.1
Werkzeug==3.0.1
wheel==0.42.0
whitenoise==6.6.0
=======
text-unidecode==1.3
toml==0.10.2
tqdm==4.66.1
trio==0.23.1
trio-websocket==0.11.1
troposphere==4.5.2
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
webdriver-manager==4.0.1
Werkzeug==3.0.1
whitenoise==6.5.0
>>>>>>> 466e9623268ee4875868b6b8d9ca6a67e70f5d8a
=======
trio-websocket==0.11.1
troposphere==4.6.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.2.0
virtualenv==20.25.0
wcwidth==0.2.13
webdriver-manager==4.0.1
Werkzeug==3.0.1
wheel==0.42.0
whitenoise==6.6.0
>>>>>>> 1e155a63d7b775eeaf726c2d137dd5d6d9ef8a48
wsproto==1.2.0
zappa==0.58.0
